perf: optimize hot paths for mobile GPU performance
- Replace aabbTransform 8-corner method with Arvo's fast AABB transform
  (eliminates 8 matrix-vector multiplies per mesh per frame during culling)
- Radix sort: ping-pong between index arrays instead of copy-back per pass
  (eliminates 4×N element copies per frame)
- Use squared distance in sort keys to avoid per-mesh Math.sqrt
- Pre-allocate cascade frustum corners Float32Array (both renderers)
  to eliminate 3 array allocations per frame
- Cache Math.tan(fovY/2) in cascade computation (both renderers)
- Replace fill(0) in mat4Identity/mat4Perspective/mat4Ortho/mat4OrthoZO
  with direct element assignment (avoids memset + overwrite pattern)
- Use Int32Array for animation cached keyframe indices
- Replace ** exponentiation with multiply in easeInOutQuad and lookAt
- Use mediump precision in basic and shadow depth fragment shaders
- Add auto-growing sort state capacity to prevent silent mesh drops

+++ src/renderer/shaders.ts
@@ -138,7 +138,7 @@ void main() {
 `
 
 export const SHADOW_DEPTH_FRAG = `#version 300 es
-precision highp float;
+precision mediump float; // Empty fragment shader — no precision needed
 void main() {}
 `
 
@@ -376,7 +376,7 @@ void main() {
 `
 
 export const BASIC_FRAG = `#version 300 es
-precision highp float;
+precision mediump float; // No precision-sensitive ops — mediump saves ALU registers on mobile GPUs
 
 in vec2 v_uv;
